refactor(v2-ontology): move things into subpackage (ongoing)
