+++ webapi/src/main/scala/org/knora/webapi/responders/v2/OntologyResponderV2.scala
@@ -17,33 +17,37 @@
  * License along with Knora.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.knora.webapi
-package responders.v2
+package org.knora.webapi.responders.v2
 
-import exceptions._
-import feature.FeatureFactoryConfig
-import messages.IriConversions._
-import messages.StringFormatter.{SalsahGuiAttribute, SalsahGuiAttributeDefinition}
-import messages.admin.responder.projectsmessages.{ProjectGetRequestADM, ProjectGetResponseADM, ProjectIdentifierADM}
-import messages.admin.responder.usersmessages.UserADM
-import messages.store.triplestoremessages._
-import messages.util.rdf.{SparqlSelectResult, VariableResultsRow}
-import messages.util.{ErrorHandlingMap, KnoraSystemInstances, OntologyUtil, ResponderData}
-import messages.v2.responder.{CanDoResponseV2, SuccessResponseV2}
-import messages.v2.responder.ontologymessages.Cardinality.KnoraCardinalityInfo
-import messages.v2.responder.ontologymessages._
-import messages.v2.responder.standoffmessages.StandoffDataTypeClasses
-import messages.{OntologyConstants, SmartIri}
-import responders.Responder.handleUnexpectedMessage
-import responders.{IriLocker, Responder}
 import util._
 import util.cache.CacheUtil
-
 import akka.http.scaladsl.util.FastFuture
 import akka.pattern._
-import org.knora.webapi.responders.v2.ontology.Cache.{ONTOLOGY_CACHE_LOCK_IRI, OntologyCacheData, makeOntologyCache}
-import org.knora.webapi.responders.v2.ontology.{Cache, DeleteCardinalitiesFromClass}
-import org.knora.webapi.responders.v2.ontology.Ontology.OntologyGraph
+import org.knora.webapi.messages.util.ResponderData
+import org.knora.webapi.messages.v2.responder.ontologymessages.{
+  CheckSubClassRequestV2,
+  ClassesGetRequestV2,
+  EntityInfoGetRequestV2,
+  LoadOntologiesRequestV2,
+  OntologiesResponderRequestV2,
+  OntologyEntitiesGetRequestV2,
+  OntologyKnoraEntityIrisGetRequestV2,
+  OntologyMetadataGetByProjectRequestV2,
+  PropertiesGetRequestV2,
+  StandoffAllPropertyEntitiesGetRequestV2,
+  StandoffClassesWithDataTypeGetRequestV2,
+  StandoffEntityInfoGetRequestV2,
+  SubClassesGetRequestV2
+}
+import org.knora.webapi.responders.Responder
+import org.knora.webapi.responders.v2.ontology.Cache.{
+  ONTOLOGY_CACHE_LOCK_IRI,
+  OntologyCacheData,
+  getCacheData,
+  makeOntologyCache
+}
+import org.knora.webapi.responders.v2.ontology.{Cache, DeleteCardinalitiesFromClass, Ontology}
+import org.knora.webapi.responders.v2.ontology.Ontology.{OntologyGraph, checkCardinalitiesBeforeAdding}
 
 import java.time.Instant
 import scala.collection.immutable
@@ -128,31 +132,6 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
   }
 
   /**
-    * Updates the ontology cache.
-    *
-    * @param cacheData the updated data to be cached.
-    */
-  private def storeCacheData(cacheData: OntologyCacheData): Unit = {
-    CacheUtil.put(cacheName = OntologyCacheName, key = OntologyCacheKey, value = cacheData)
-  }
-
-  /**
-    * Gets the ontology data from the cache.
-    *
-    * @return an [[OntologyCacheData]]
-    */
-  private def getCacheData: Future[OntologyCacheData] = {
-    Future {
-      CacheUtil.get[OntologyCacheData](cacheName = OntologyCacheName, key = OntologyCacheKey) match {
-        case Some(data) => data
-        case None =>
-          throw ApplicationCacheException(
-            s"The Knora API server has not loaded any ontologies, perhaps because of an invalid ontology")
-      }
-    }
-  }
-
-  /**
     * Given a list of resource IRIs and a list of property IRIs (ontology entities), returns an [[EntityInfoGetResponseV2]] describing both resource and property entities.
     *
     * @param classIris      the IRIs of the resource entities to be queried.
@@ -219,7 +198,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
                 val internalPropertyIri = propertyIri.toOntologySchema(InternalSchema)
 
                 // If it's a link value property and it's requested in the simple schema, it's unavailable.
-                if (apiV2Schema == ApiV2Simple && isLinkValueProp(internalPropertyIri, cacheData)) {
+                if (apiV2Schema == ApiV2Simple && Ontology.isLinkValueProp(internalPropertyIri, cacheData)) {
                   acc + propertyIri
                 } else {
                   // Is it among the properties removed from the internal ontology in the requested schema?
@@ -505,7 +484,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
       OntologyKnoraEntitiesIriInfoV2(
         ontologyIri = ontologyIri,
         propertyIris = ontology.properties.keySet.filter { propertyIri =>
-          isKnoraResourceProperty(propertyIri, cacheData)
+          Ontology.isKnoraResourceProperty(propertyIri, cacheData)
         },
         classIris = ontology.classes.filter {
           case (_, classDef) => classDef.isResourceClass
@@ -701,120 +680,6 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
   }
 
   /**
-    * Reads an ontology's metadata.
-    *
-    * @param internalOntologyIri  the ontology's internal IRI.
-    * @param featureFactoryConfig the feature factory configuration.
-    * @return an [[OntologyMetadataV2]], or [[None]] if the ontology is not found.
-    */
-  private def loadOntologyMetadata(internalOntologyIri: SmartIri,
-                                   featureFactoryConfig: FeatureFactoryConfig): Future[Option[OntologyMetadataV2]] = {
-    for {
-      _ <- Future {
-        if (!internalOntologyIri.getOntologySchema.contains(InternalSchema)) {
-          throw AssertionException(s"Expected an internal ontology IRI: $internalOntologyIri")
-        }
-      }
-
-      getOntologyInfoSparql = org.knora.webapi.messages.twirl.queries.sparql.v2.txt
-        .getOntologyInfo(
-          triplestore = settings.triplestoreType,
-          ontologyIri = internalOntologyIri
-        )
-        .toString()
-
-      getOntologyInfoResponse <- (storeManager ? SparqlConstructRequest(
-        sparql = getOntologyInfoSparql,
-        featureFactoryConfig = featureFactoryConfig
-      )).mapTo[SparqlConstructResponse]
-
-      metadata: Option[OntologyMetadataV2] = if (getOntologyInfoResponse.statements.isEmpty) {
-        None
-      } else {
-        getOntologyInfoResponse.statements.get(internalOntologyIri.toString) match {
-          case Some(statements: Seq[(IRI, String)]) =>
-            val statementMap: Map[IRI, Seq[String]] = statements
-              .groupBy {
-                case (pred, _) => pred
-              }
-              .map {
-                case (pred, predStatements) =>
-                  pred -> predStatements.map {
-                    case (_, obj) => obj
-                  }
-              }
-
-            val projectIris: Seq[String] = statementMap.getOrElse(
-              OntologyConstants.KnoraBase.AttachedToProject,
-              throw InconsistentRepositoryDataException(
-                s"Ontology $internalOntologyIri has no knora-base:attachedToProject")
-            )
-            val labels: Seq[String] = statementMap.getOrElse(OntologyConstants.Rdfs.Label, Seq.empty[String])
-            val comments: Seq[String] = statementMap.getOrElse(OntologyConstants.Rdfs.Comment, Seq.empty[String])
-            val lastModDates: Seq[String] =
-              statementMap.getOrElse(OntologyConstants.KnoraBase.LastModificationDate, Seq.empty[String])
-
-            val projectIri = if (projectIris.size > 1) {
-              throw InconsistentRepositoryDataException(
-                s"Ontology $internalOntologyIri has more than one knora-base:attachedToProject")
-            } else {
-              projectIris.head.toSmartIri
-            }
-
-            if (!internalOntologyIri.isKnoraBuiltInDefinitionIri) {
-              if (projectIri.toString == OntologyConstants.KnoraAdmin.SystemProject) {
-                throw InconsistentRepositoryDataException(
-                  s"Ontology $internalOntologyIri cannot be in project ${OntologyConstants.KnoraAdmin.SystemProject}")
-              }
-
-              if (internalOntologyIri.isKnoraSharedDefinitionIri && projectIri.toString != OntologyConstants.KnoraAdmin.DefaultSharedOntologiesProject) {
-                throw InconsistentRepositoryDataException(
-                  s"Shared ontology $internalOntologyIri must be in project ${OntologyConstants.KnoraAdmin.DefaultSharedOntologiesProject}")
-              }
-            }
-
-            val label: String = if (labels.size > 1) {
-              throw InconsistentRepositoryDataException(s"Ontology $internalOntologyIri has more than one rdfs:label")
-            } else if (labels.isEmpty) {
-              internalOntologyIri.getOntologyName
-            } else {
-              labels.head
-            }
-
-            val comment: Option[String] = if (comments.size > 1) {
-              throw InconsistentRepositoryDataException(s"Ontology $internalOntologyIri has more than one rdfs:comment")
-            } else comments.headOption
-
-            val lastModificationDate: Option[Instant] = if (lastModDates.size > 1) {
-              throw InconsistentRepositoryDataException(
-                s"Ontology $internalOntologyIri has more than one ${OntologyConstants.KnoraBase.LastModificationDate}")
-            } else if (lastModDates.isEmpty) {
-              None
-            } else {
-              val dateStr = lastModDates.head
-              Some(
-                stringFormatter.xsdDateTimeStampToInstant(
-                  dateStr,
-                  throw InconsistentRepositoryDataException(
-                    s"Invalid ${OntologyConstants.KnoraBase.LastModificationDate}: $dateStr")))
-            }
-
-            Some(
-              OntologyMetadataV2(
-                ontologyIri = internalOntologyIri,
-                projectIri = Some(projectIri),
-                label = Some(label),
-                comment = comment,
-                lastModificationDate = lastModificationDate
-              ))
-
-          case None => None
-        }
-      }
-    } yield metadata
-  }
-
-  /**
     * Creates a new, empty ontology.
     *
     * @param createOntologyRequest the request message.
@@ -823,10 +688,12 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
   private def createOntology(createOntologyRequest: CreateOntologyRequestV2): Future[ReadOntologyMetadataV2] = {
     def makeTaskFuture(internalOntologyIri: SmartIri): Future[ReadOntologyMetadataV2] = {
       for {
-        cacheData <- getCacheData
+        cacheData <- Cache.getCacheData
 
         // Make sure the ontology doesn't already exist.
-        existingOntologyMetadata: Option[OntologyMetadataV2] <- loadOntologyMetadata(
+        existingOntologyMetadata: Option[OntologyMetadataV2] <- Ontology.loadOntologyMetadata(
+          settings,
+          storeManager,
           internalOntologyIri = internalOntologyIri,
           featureFactoryConfig = createOntologyRequest.featureFactoryConfig
         )
@@ -893,7 +760,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
 
         // Update the ontology cache with the unescaped metadata.
 
-        _ = storeCacheData(
+        _ = Cache.storeCacheData(
           cacheData.copy(
             ontologies = cacheData.ontologies + (internalOntologyIri -> ReadOntologyV2(
               ontologyMetadata = unescapedNewMetadata))
@@ -953,13 +820,13 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
         cacheData <- getCacheData
 
         // Check that the user has permission to update the ontology.
-        projectIri <- checkPermissionsForOntologyUpdate(
+        projectIri <- Ontology.checkPermissionsForOntologyUpdate(
           internalOntologyIri = internalOntologyIri,
           requestingUser = changeOntologyMetadataRequest.requestingUser
         )
 
         // Check that the ontology exists and has not been updated by another user since the client last read its metadata.
-        _ <- checkOntologyLastModificationDateBeforeUpdate(
+        _ <- Ontology.checkOntologyLastModificationDateBeforeUpdate(
           internalOntologyIri = internalOntologyIri,
           expectedLastModificationDate = changeOntologyMetadataRequest.lastModificationDate,
           featureFactoryConfig = changeOntologyMetadataRequest.featureFactoryConfig
@@ -1034,7 +901,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
 
         // Update the ontology cache with the unescaped metadata.
 
-        _ = storeCacheData(
+        _ = Cache.storeCacheData(
           cacheData.copy(
             ontologies = cacheData.ontologies + (internalOntologyIri -> cacheData
               .ontologies(internalOntologyIri)
@@ -1045,7 +912,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
     }
 
     for {
-      _ <- checkExternalOntologyIriForUpdate(changeOntologyMetadataRequest.ontologyIri)
+      _ <- Ontology.checkExternalOntologyIriForUpdate(changeOntologyMetadataRequest.ontologyIri)
       internalOntologyIri = changeOntologyMetadataRequest.ontologyIri.toOntologySchema(InternalSchema)
 
       // Do the remaining pre-update checks and the update while holding a global ontology cache lock.
@@ -1064,13 +931,13 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
         cacheData <- getCacheData
 
         // Check that the user has permission to update the ontology.
-        projectIri <- checkPermissionsForOntologyUpdate(
+        projectIri <- Ontology.checkPermissionsForOntologyUpdate(
           internalOntologyIri = internalOntologyIri,
           requestingUser = deleteOntologyCommentRequestV2.requestingUser
         )
 
         // Check that the ontology exists and has not been updated by another user since the client last read its metadata.
-        _ <- checkOntologyLastModificationDateBeforeUpdate(
+        _ <- Ontology.checkOntologyLastModificationDateBeforeUpdate(
           internalOntologyIri = internalOntologyIri,
           expectedLastModificationDate = deleteOntologyCommentRequestV2.lastModificationDate,
           featureFactoryConfig = deleteOntologyCommentRequestV2.featureFactoryConfig
@@ -1127,7 +994,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
 
         // Update the ontology cache with the unescaped metadata.
 
-        _ = storeCacheData(
+        _ = Cache.storeCacheData(
           cacheData.copy(
             ontologies = cacheData.ontologies + (internalOntologyIri -> cacheData
               .ontologies(internalOntologyIri)
@@ -1138,7 +1005,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
     }
 
     for {
-      _ <- checkExternalOntologyIriForUpdate(deleteOntologyCommentRequestV2.ontologyIri)
+      _ <- Ontology.checkExternalOntologyIriForUpdate(deleteOntologyCommentRequestV2.ontologyIri)
       internalOntologyIri = deleteOntologyCommentRequestV2.ontologyIri.toOntologySchema(InternalSchema)
 
       // Do the remaining pre-update checks and the update while holding a global ontology cache lock.
@@ -1163,7 +1030,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
         internalClassDef: ClassInfoContentV2 = createClassRequest.classInfoContent.toOntologySchema(InternalSchema)
 
         // Check that the ontology exists and has not been updated by another user since the client last read it.
-        _ <- checkOntologyLastModificationDateBeforeUpdate(
+        _ <- Ontology.checkOntologyLastModificationDateBeforeUpdate(
           internalOntologyIri = internalOntologyIri,
           expectedLastModificationDate = createClassRequest.lastModificationDate,
           featureFactoryConfig = createClassRequest.featureFactoryConfig
@@ -1194,7 +1061,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
 
         missingBaseClasses = internalClassDef.subClassOf
           .filter(_.isKnoraInternalEntityIri)
-          .filter(baseClassIri => !isKnoraInternalResourceClass(baseClassIri, cacheData))
+          .filter(baseClassIri => !Ontology.isKnoraInternalResourceClass(baseClassIri, cacheData))
 
         _ = if (missingBaseClasses.nonEmpty) {
           throw BadRequestException(
@@ -1222,14 +1089,15 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
         }
 
         // Check that the cardinalities are valid, and add any inherited cardinalities.
-        (internalClassDefWithLinkValueProps, cardinalitiesForClassWithInheritance) = checkCardinalitiesBeforeAdding(
-          internalClassDef = internalClassDef,
-          allBaseClassIris = allBaseClassIris.toSet,
-          cacheData = cacheData
-        )
+        (internalClassDefWithLinkValueProps, cardinalitiesForClassWithInheritance) = Ontology
+          .checkCardinalitiesBeforeAdding(
+            internalClassDef = internalClassDef,
+            allBaseClassIris = allBaseClassIris.toSet,
+            cacheData = cacheData
+          )
 
         // Check that the class definition doesn't refer to any non-shared ontologies in other projects.
-        _ = checkOntologyReferencesInClassDef(
+        _ = Cache.checkOntologyReferencesInClassDef(
           ontologyCacheData = cacheData,
           classDef = internalClassDefWithLinkValueProps,
           errorFun = { msg: String =>
@@ -1254,13 +1122,13 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
           canBeInstantiated = true,
           inheritedCardinalities = inheritedCardinalities,
           knoraResourceProperties = propertyIrisOfAllCardinalitiesForClass.filter(propertyIri =>
-            isKnoraResourceProperty(propertyIri, cacheData)),
+            Ontology.isKnoraResourceProperty(propertyIri, cacheData)),
           linkProperties =
-            propertyIrisOfAllCardinalitiesForClass.filter(propertyIri => isLinkProp(propertyIri, cacheData)),
-          linkValueProperties =
-            propertyIrisOfAllCardinalitiesForClass.filter(propertyIri => isLinkValueProp(propertyIri, cacheData)),
-          fileValueProperties =
-            propertyIrisOfAllCardinalitiesForClass.filter(propertyIri => isFileValueProp(propertyIri, cacheData))
+            propertyIrisOfAllCardinalitiesForClass.filter(propertyIri => Ontology.isLinkProp(propertyIri, cacheData)),
+          linkValueProperties = propertyIrisOfAllCardinalitiesForClass.filter(propertyIri =>
+            Ontology.isLinkValueProp(propertyIri, cacheData)),
+          fileValueProperties = propertyIrisOfAllCardinalitiesForClass.filter(propertyIri =>
+            Ontology.isFileValueProp(propertyIri, cacheData))
         )
 
         // Add the SPARQL-escaped class to the triplestore.
@@ -1282,7 +1150,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
 
         // Check that the ontology's last modification date was updated.
 
-        _ <- checkOntologyLastModificationDateAfterUpdate(
+        _ <- Ontology.checkOntologyLastModificationDateAfterUpdate(
           internalOntologyIri = internalOntologyIri,
           expectedLastModificationDate = currentTime,
           featureFactoryConfig = createClassRequest.featureFactoryConfig
@@ -1290,7 +1158,9 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
 
         // Check that the data that was saved corresponds to the data that was submitted.
 
-        loadedClassDef <- loadClassDefinition(
+        loadedClassDef <- Ontology.loadClassDefinition(
+          settings,
+          storeManager,
           classIri = internalClassIri,
           featureFactoryConfig = createClassRequest.featureFactoryConfig
         )
@@ -1303,7 +1173,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
         // Update the cache.
 
         updatedSubClassOfRelations = cacheData.subClassOfRelations + (internalClassIri -> allBaseClassIris)
-        updatedSuperClassOfRelations = calculateSuperClassOfRelations(updatedSubClassOfRelations)
+        updatedSuperClassOfRelations = Ontology.calculateSuperClassOfRelations(updatedSubClassOfRelations)
 
         updatedOntology = ontology.copy(
           ontologyMetadata = ontology.ontologyMetadata.copy(
@@ -1312,7 +1182,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
           classes = ontology.classes + (internalClassIri -> readClassInfo)
         )
 
-        _ = storeCacheData(
+        _ = Cache.storeCacheData(
           cacheData.copy(
             ontologies = cacheData.ontologies + (internalOntologyIri -> updatedOntology),
             subClassOfRelations = updatedSubClassOfRelations,
@@ -1335,7 +1205,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
       externalClassIri = createClassRequest.classInfoContent.classIri
       externalOntologyIri = externalClassIri.getOntologyFromEntity
 
-      _ <- checkOntologyAndEntityIrisForUpdate(
+      _ <- Ontology.checkOntologyAndEntityIrisForUpdate(
         externalOntologyIri = externalOntologyIri,
         externalEntityIri = externalClassIri,
         requestingUser = requestingUser
@@ -1358,279 +1228,6 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
   }
 
   /**
-    * Before creating a new class or adding cardinalities to an existing class, checks the validity of the
-    * cardinalities directly defined on the class. Adds link value properties for the corresponding
-    * link properties.
-    *
-    * @param internalClassDef        the internal definition of the class.
-    * @param allBaseClassIris        the IRIs of all the class's base classes, including the class itself.
-    * @param cacheData               the ontology cache.
-    * @param existingLinkPropsToKeep the link properties that are already defined on the class and that
-    *                                will be kept after the update.
-    * @return the updated class definition, and the cardinalities resulting from inheritance.
-    */
-  private def checkCardinalitiesBeforeAdding(
-      internalClassDef: ClassInfoContentV2,
-      allBaseClassIris: Set[SmartIri],
-      cacheData: OntologyCacheData,
-      existingLinkPropsToKeep: Set[SmartIri] = Set.empty): (ClassInfoContentV2, Map[SmartIri, KnoraCardinalityInfo]) = {
-    // If the class has cardinalities, check that the properties are already defined as Knora properties.
-
-    val propertyDefsForDirectCardinalities: Set[ReadPropertyInfoV2] = internalClassDef.directCardinalities.keySet.map {
-      propertyIri =>
-        if (!isKnoraResourceProperty(propertyIri, cacheData) || propertyIri.toString == OntologyConstants.KnoraBase.ResourceProperty || propertyIri.toString == OntologyConstants.KnoraBase.HasValue) {
-          throw NotFoundException(s"Invalid property for cardinality: <${propertyIri.toOntologySchema(ApiV2Complex)}>")
-        }
-
-        cacheData.ontologies(propertyIri.getOntologyFromEntity).properties(propertyIri)
-    }
-
-    val existingLinkValuePropsToKeep = existingLinkPropsToKeep.map(_.fromLinkPropToLinkValueProp)
-    val newLinkPropsInClass: Set[SmartIri] = propertyDefsForDirectCardinalities
-      .filter(_.isLinkProp)
-      .map(_.entityInfoContent.propertyIri) -- existingLinkValuePropsToKeep
-    val newLinkValuePropsInClass: Set[SmartIri] = propertyDefsForDirectCardinalities
-      .filter(_.isLinkValueProp)
-      .map(_.entityInfoContent.propertyIri) -- existingLinkValuePropsToKeep
-
-    // Don't allow link value prop cardinalities to be included in the request.
-
-    if (newLinkValuePropsInClass.nonEmpty) {
-      throw BadRequestException(s"In class ${internalClassDef.classIri.toOntologySchema(ApiV2Complex)}, cardinalities have been submitted for one or more link value properties: ${newLinkValuePropsInClass
-        .map(_.toOntologySchema(ApiV2Complex))
-        .mkString(", ")}. Just submit the link properties, and the link value properties will be included automatically.")
-    }
-
-    // Add a link value prop cardinality for each new link prop cardinality.
-
-    val linkValuePropCardinalitiesToAdd: Map[SmartIri, KnoraCardinalityInfo] = newLinkPropsInClass.map { linkPropIri =>
-      val linkValuePropIri = linkPropIri.fromLinkPropToLinkValueProp
-
-      // Ensure that the link value prop exists.
-      cacheData
-        .ontologies(linkValuePropIri.getOntologyFromEntity)
-        .properties
-        .getOrElse(linkValuePropIri, throw NotFoundException(s"Link value property <$linkValuePropIri> not found"))
-
-      linkValuePropIri -> internalClassDef.directCardinalities(linkPropIri)
-    }.toMap
-
-    val classDefWithAddedLinkValueProps: ClassInfoContentV2 = internalClassDef.copy(
-      directCardinalities = internalClassDef.directCardinalities ++ linkValuePropCardinalitiesToAdd
-    )
-
-    // Get the cardinalities that the class can inherit.
-
-    val cardinalitiesAvailableToInherit: Map[SmartIri, KnoraCardinalityInfo] =
-      classDefWithAddedLinkValueProps.subClassOf.flatMap { baseClassIri =>
-        cacheData.ontologies(baseClassIri.getOntologyFromEntity).classes(baseClassIri).allCardinalities
-      }.toMap
-
-    // Check that the cardinalities directly defined on the class are compatible with any inheritable
-    // cardinalities, and let directly-defined cardinalities override cardinalities in base classes.
-
-    val cardinalitiesForClassWithInheritance: Map[SmartIri, KnoraCardinalityInfo] = overrideCardinalities(
-      classIri = internalClassDef.classIri,
-      thisClassCardinalities = classDefWithAddedLinkValueProps.directCardinalities,
-      inheritableCardinalities = cardinalitiesAvailableToInherit,
-      allSubPropertyOfRelations = cacheData.subPropertyOfRelations,
-      errorSchema = ApiV2Complex,
-      errorFun = { msg: String =>
-        throw BadRequestException(msg)
-      }
-    )
-
-    // Check that the class is a subclass of all the classes that are subject class constraints of the Knora resource properties in its cardinalities.
-
-    val knoraResourcePropertyIrisInCardinalities = cardinalitiesForClassWithInheritance.keySet.filter { propertyIri =>
-      isKnoraResourceProperty(
-        propertyIri = propertyIri,
-        cacheData = cacheData
-      )
-    }
-
-    val allClassCardinalityKnoraPropertyDefs: Map[SmartIri, PropertyInfoContentV2] =
-      knoraResourcePropertyIrisInCardinalities.map { propertyIri =>
-        propertyIri -> cacheData.ontologies(propertyIri.getOntologyFromEntity).properties(propertyIri).entityInfoContent
-      }.toMap
-
-    checkSubjectClassConstraintsViaCardinalities(
-      internalClassDef = classDefWithAddedLinkValueProps,
-      allBaseClassIris = allBaseClassIris,
-      allClassCardinalityKnoraPropertyDefs = allClassCardinalityKnoraPropertyDefs,
-      errorSchema = ApiV2Complex,
-      errorFun = { msg: String =>
-        throw BadRequestException(msg)
-      }
-    )
-
-    // It cannot have cardinalities both on property P and on a subproperty of P.
-
-    val maybePropertyAndSubproperty: Option[(SmartIri, SmartIri)] = findPropertyAndSubproperty(
-      propertyIris = cardinalitiesForClassWithInheritance.keySet,
-      subPropertyOfRelations = cacheData.subPropertyOfRelations
-    )
-
-    maybePropertyAndSubproperty match {
-      case Some((basePropertyIri, propertyIri)) =>
-        throw BadRequestException(
-          s"Class <${classDefWithAddedLinkValueProps.classIri.toOntologySchema(ApiV2Complex)}> has a cardinality on property <${basePropertyIri
-            .toOntologySchema(ApiV2Complex)}> and on its subproperty <${propertyIri.toOntologySchema(ApiV2Complex)}>")
-
-      case None => ()
-    }
-
-    // Check for invalid cardinalities on boolean properties.
-    checkForInvalidBooleanCardinalities(
-      classIri = internalClassDef.classIri,
-      directCardinalities = internalClassDef.directCardinalities,
-      allPropertyDefs = cacheData.allPropertyDefs,
-      schemaForErrors = ApiV2Complex,
-      errorFun = { msg: String =>
-        throw BadRequestException(msg)
-      }
-    )
-
-    (classDefWithAddedLinkValueProps, cardinalitiesForClassWithInheritance)
-  }
-
-  /**
-    * Given the IRI of a base class, updates inherited cardinalities in subclasses.
-    *
-    * @param baseClassIri the internal IRI of the base class.
-    * @param cacheData the ontology cache.
-    *
-    * @return the updated ontology cache.
-    */
-  private def updateSubClasses(baseClassIri: SmartIri, cacheData: OntologyCacheData): OntologyCacheData = {
-    // Get the class definitions of all the subclasses of the base class.
-
-    val allSubClassIris: Set[SmartIri] = cacheData.superClassOfRelations(baseClassIri)
-
-    val allSubClasses: Set[ReadClassInfoV2] = allSubClassIris.map { subClassIri =>
-      cacheData.ontologies(subClassIri.getOntologyFromEntity).classes(subClassIri)
-    }
-
-    // Filter them to get only the direct subclasses.
-
-    val directSubClasses: Set[ReadClassInfoV2] = allSubClasses.filter { subClass =>
-      subClass.entityInfoContent.subClassOf
-        .contains(baseClassIri) && subClass.entityInfoContent.classIri != baseClassIri
-    }
-
-    // Iterate over the subclasses, updating cardinalities.
-    val cacheDataWithUpdatedSubClasses = directSubClasses.foldLeft(cacheData) {
-      case (cacheDataAcc: OntologyCacheData, directSubClass: ReadClassInfoV2) =>
-        val directSubClassIri = directSubClass.entityInfoContent.classIri
-
-        // Get the cardinalities that this subclass can inherit from its direct base classes.
-
-        val inheritableCardinalities: Map[SmartIri, KnoraCardinalityInfo] =
-          directSubClass.entityInfoContent.subClassOf.flatMap { baseClassIri =>
-            cacheData.ontologies(baseClassIri.getOntologyFromEntity).classes(baseClassIri).allCardinalities
-          }.toMap
-
-        // Override inherited cardinalities with directly defined cardinalities.
-        val newInheritedCardinalities: Map[SmartIri, KnoraCardinalityInfo] = overrideCardinalities(
-          classIri = directSubClassIri,
-          thisClassCardinalities = directSubClass.entityInfoContent.directCardinalities,
-          inheritableCardinalities = inheritableCardinalities,
-          allSubPropertyOfRelations = cacheData.subPropertyOfRelations,
-          errorSchema = ApiV2Complex,
-          errorFun = { msg: String =>
-            throw BadRequestException(msg)
-          }
-        )
-
-        // Update the cache.
-
-        val ontologyIri = directSubClass.entityInfoContent.classIri.getOntologyFromEntity
-        val ontology: ReadOntologyV2 = cacheDataAcc.ontologies(ontologyIri)
-
-        val updatedOntology = ontology.copy(
-          classes = ontology.classes + (directSubClassIri -> directSubClass.copy(
-            inheritedCardinalities = newInheritedCardinalities
-          ))
-        )
-
-        cacheDataAcc.copy(
-          ontologies = cacheDataAcc.ontologies + (ontologyIri -> updatedOntology)
-        )
-    }
-
-    // Recurse to subclasses of subclasses.
-
-    directSubClasses.map(_.entityInfoContent.classIri).foldLeft(cacheDataWithUpdatedSubClasses) {
-      case (cacheDataAcc: OntologyCacheData, directSubClassIri: SmartIri) =>
-        updateSubClasses(baseClassIri = directSubClassIri, cacheDataAcc)
-    }
-  }
-
-  /**
-    * Given a set of property IRIs, determines whether the set contains a property P and a subproperty of P.
-    *
-    * @param propertyIris           the set of property IRIs.
-    * @param subPropertyOfRelations all the subproperty relations in the triplestore.
-    * @return a property and its subproperty, if found.
-    */
-  private def findPropertyAndSubproperty(
-      propertyIris: Set[SmartIri],
-      subPropertyOfRelations: Map[SmartIri, Set[SmartIri]]): Option[(SmartIri, SmartIri)] = {
-    propertyIris.flatMap { propertyIri =>
-      val maybeBasePropertyIri: Option[SmartIri] = (propertyIris - propertyIri).find { otherPropertyIri =>
-        subPropertyOfRelations.get(propertyIri).exists { baseProperties: Set[SmartIri] =>
-          baseProperties.contains(otherPropertyIri)
-        }
-      }
-
-      maybeBasePropertyIri.map { basePropertyIri =>
-        (basePropertyIri, propertyIri)
-      }
-    }.headOption
-  }
-
-  /**
-    * Checks that a class is a subclass of all the classes that are subject class constraints of the Knora resource properties in its cardinalities.
-    *
-    * @param internalClassDef                     the class definition.
-    * @param allBaseClassIris                     the IRIs of all the class's base classes.
-    * @param allClassCardinalityKnoraPropertyDefs the definitions of all the Knora resource properties on which the class has cardinalities (whether directly defined
-    *                                             or inherited).
-    * @param errorSchema                          the ontology schema to be used in error messages.
-    * @param errorFun                             a function that throws an exception. It will be called with an error message argument if the cardinalities are invalid.
-    */
-  private def checkSubjectClassConstraintsViaCardinalities(
-      internalClassDef: ClassInfoContentV2,
-      allBaseClassIris: Set[SmartIri],
-      allClassCardinalityKnoraPropertyDefs: Map[SmartIri, PropertyInfoContentV2],
-      errorSchema: OntologySchema,
-      errorFun: String => Nothing): Unit = {
-    allClassCardinalityKnoraPropertyDefs.foreach {
-      case (propertyIri, propertyDef) =>
-        propertyDef.predicates.get(OntologyConstants.KnoraBase.SubjectClassConstraint.toSmartIri) match {
-          case Some(subjectClassConstraintPred) =>
-            val subjectClassConstraint = subjectClassConstraintPred.requireIriObject(
-              throw InconsistentRepositoryDataException(
-                s"Property $propertyIri has an invalid object for ${OntologyConstants.KnoraBase.SubjectClassConstraint}"))
-
-            if (!allBaseClassIris.contains(subjectClassConstraint)) {
-              val hasOrWouldInherit = if (internalClassDef.directCardinalities.contains(propertyIri)) {
-                "has"
-              } else {
-                "would inherit"
-              }
-
-              errorFun(s"Class ${internalClassDef.classIri.toOntologySchema(errorSchema)} $hasOrWouldInherit a cardinality for property ${propertyIri
-                .toOntologySchema(errorSchema)}, but is not a subclass of that property's ${OntologyConstants.KnoraBase.SubjectClassConstraint.toSmartIri
-                .toOntologySchema(errorSchema)}, ${subjectClassConstraint.toOntologySchema(errorSchema)}")
-            }
-
-          case None => ()
-        }
-    }
-  }
-
-  /**
     * Adds cardinalities to an existing class definition.
     *
     * @param addCardinalitiesRequest the request to add the cardinalities.
@@ -1644,7 +1241,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
         internalClassDef: ClassInfoContentV2 = addCardinalitiesRequest.classInfoContent.toOntologySchema(InternalSchema)
 
         // Check that the ontology exists and has not been updated by another user since the client last read it.
-        _ <- checkOntologyLastModificationDateBeforeUpdate(
+        _ <- Ontology.checkOntologyLastModificationDateBeforeUpdate(
           internalOntologyIri = internalOntologyIri,
           expectedLastModificationDate = addCardinalitiesRequest.lastModificationDate,
           featureFactoryConfig = addCardinalitiesRequest.featureFactoryConfig
@@ -1726,7 +1323,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
         )
 
         // Check that the class definition doesn't refer to any non-shared ontologies in other projects.
-        _ = checkOntologyReferencesInClassDef(
+        _ = Cache.checkOntologyReferencesInClassDef(
           ontologyCacheData = cacheData,
           classDef = newInternalClassDefWithLinkValueProps,
           errorFun = { msg: String =>
@@ -1750,13 +1347,13 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
           canBeInstantiated = true,
           inheritedCardinalities = inheritedCardinalities,
           knoraResourceProperties = propertyIrisOfAllCardinalitiesForClass.filter(propertyIri =>
-            isKnoraResourceProperty(propertyIri, cacheData)),
+            Ontology.isKnoraResourceProperty(propertyIri, cacheData)),
           linkProperties =
-            propertyIrisOfAllCardinalitiesForClass.filter(propertyIri => isLinkProp(propertyIri, cacheData)),
-          linkValueProperties =
-            propertyIrisOfAllCardinalitiesForClass.filter(propertyIri => isLinkValueProp(propertyIri, cacheData)),
-          fileValueProperties =
-            propertyIrisOfAllCardinalitiesForClass.filter(propertyIri => isFileValueProp(propertyIri, cacheData))
+            propertyIrisOfAllCardinalitiesForClass.filter(propertyIri => Ontology.isLinkProp(propertyIri, cacheData)),
+          linkValueProperties = propertyIrisOfAllCardinalitiesForClass.filter(propertyIri =>
+            Ontology.isLinkValueProp(propertyIri, cacheData)),
+          fileValueProperties = propertyIrisOfAllCardinalitiesForClass.filter(propertyIri =>
+            Ontology.isFileValueProp(propertyIri, cacheData))
         )
 
         // Add the cardinalities to the class definition in the triplestore.
@@ -1779,7 +1376,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
 
         // Check that the ontology's last modification date was updated.
 
-        _ <- checkOntologyLastModificationDateAfterUpdate(
+        _ <- Ontology.checkOntologyLastModificationDateAfterUpdate(
           internalOntologyIri = internalOntologyIri,
           expectedLastModificationDate = currentTime,
           featureFactoryConfig = addCardinalitiesRequest.featureFactoryConfig
@@ -1787,7 +1384,9 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
 
         // Check that the data that was saved corresponds to the data that was submitted.
 
-        loadedClassDef <- loadClassDefinition(
+        loadedClassDef <- Ontology.loadClassDefinition(
+          settings,
+          storeManager,
           classIri = internalClassIri,
           featureFactoryConfig = addCardinalitiesRequest.featureFactoryConfig
         )
@@ -1806,11 +1405,11 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
           classes = ontology.classes + (internalClassIri -> readClassInfo)
         )
 
-        _ = storeCacheData(
-          updateSubClasses(baseClassIri = internalClassIri,
-                           cacheData = cacheData.copy(
-                             ontologies = cacheData.ontologies + (internalOntologyIri -> updatedOntology)
-                           )))
+        _ = Cache.storeCacheData(
+          Cache.updateSubClasses(baseClassIri = internalClassIri,
+                                 cacheData = cacheData.copy(
+                                   ontologies = cacheData.ontologies + (internalOntologyIri -> updatedOntology)
+                                 )))
 
         // Read the data back from the cache.
 
@@ -1828,7 +1427,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
       externalClassIri = addCardinalitiesRequest.classInfoContent.classIri
       externalOntologyIri = externalClassIri.getOntologyFromEntity
 
-      _ <- checkOntologyAndEntityIrisForUpdate(
+      _ <- Ontology.checkOntologyAndEntityIrisForUpdate(
         externalOntologyIri = externalOntologyIri,
         externalEntityIri = externalClassIri,
         requestingUser = requestingUser
@@ -1863,7 +1462,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
         internalClassDef: ClassInfoContentV2 = changeGuiOrderRequest.classInfoContent.toOntologySchema(InternalSchema)
 
         // Check that the ontology exists and has not been updated by another user since the client last read it.
-        _ <- checkOntologyLastModificationDateBeforeUpdate(
+        _ <- Ontology.checkOntologyLastModificationDateBeforeUpdate(
           internalOntologyIri = internalOntologyIri,
           expectedLastModificationDate = changeGuiOrderRequest.lastModificationDate,
           featureFactoryConfig = changeGuiOrderRequest.featureFactoryConfig
@@ -1947,7 +1546,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
 
         // Check that the ontology's last modification date was updated.
 
-        _ <- checkOntologyLastModificationDateAfterUpdate(
+        _ <- Ontology.checkOntologyLastModificationDateAfterUpdate(
           internalOntologyIri = internalOntologyIri,
           expectedLastModificationDate = currentTime,
           featureFactoryConfig = changeGuiOrderRequest.featureFactoryConfig
@@ -1955,7 +1554,9 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
 
         // Check that the data that was saved corresponds to the data that was submitted.
 
-        loadedClassDef: ClassInfoContentV2 <- loadClassDefinition(
+        loadedClassDef: ClassInfoContentV2 <- Ontology.loadClassDefinition(
+          settings,
+          storeManager,
           classIri = internalClassIri,
           featureFactoryConfig = changeGuiOrderRequest.featureFactoryConfig
         )
@@ -1976,11 +1577,11 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
 
         // Update subclasses and write the cache.
 
-        _ = storeCacheData(
-          updateSubClasses(baseClassIri = internalClassIri,
-                           cacheData = cacheData.copy(
-                             ontologies = cacheData.ontologies + (internalOntologyIri -> updatedOntology)
-                           )))
+        _ = Cache.storeCacheData(
+          Cache.updateSubClasses(baseClassIri = internalClassIri,
+                                 cacheData = cacheData.copy(
+                                   ontologies = cacheData.ontologies + (internalOntologyIri -> updatedOntology)
+                                 )))
 
         // Read the data back from the cache.
 
@@ -1999,7 +1600,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
       externalClassIri = changeGuiOrderRequest.classInfoContent.classIri
       externalOntologyIri = externalClassIri.getOntologyFromEntity
 
-      _ <- checkOntologyAndEntityIrisForUpdate(
+      _ <- Ontology.checkOntologyAndEntityIrisForUpdate(
         externalOntologyIri = externalOntologyIri,
         externalEntityIri = externalClassIri,
         requestingUser = requestingUser
@@ -2044,7 +1645,8 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
         throw BadRequestException(s"Class ${canChangeCardinalitiesRequest.classIri} does not exist")
       }
 
-      userCanUpdateOntology <- canUserUpdateOntology(internalOntologyIri, canChangeCardinalitiesRequest.requestingUser)
+      userCanUpdateOntology <- Ontology.canUserUpdateOntology(internalOntologyIri,
+                                                              canChangeCardinalitiesRequest.requestingUser)
 
       classIsUsed <- isEntityUsed(
         entityIri = internalClassIri,
@@ -2124,7 +1726,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
         )
 
         // Check that the class definition doesn't refer to any non-shared ontologies in other projects.
-        _ = checkOntologyReferencesInClassDef(
+        _ = Cache.checkOntologyReferencesInClassDef(
           ontologyCacheData = cacheData,
           classDef = newInternalClassDefWithLinkValueProps,
           errorFun = { msg: String =>
@@ -2204,7 +1806,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
           classes = ontology.classes + (internalClassIri -> readClassInfo)
         )
 
-        _ = storeCacheData(
+        _ = Cache.storeCacheData(
           cacheData.copy(
             ontologies = cacheData.ontologies + (internalOntologyIri -> updatedOntology)
           ))
@@ -2384,7 +1986,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
 
         updatedSuperClassOfRelations = calculateSuperClassOfRelations(updatedSubClassOfRelations)
 
-        _ = storeCacheData(
+        _ = Cache.storeCacheData(
           cacheData.copy(
             ontologies = cacheData.ontologies + (internalOntologyIri -> updatedOntology),
             subClassOfRelations = updatedSubClassOfRelations,
@@ -2550,7 +2152,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
           case (subProperty, baseProperties) => subProperty -> (baseProperties -- propertiesToRemoveFromCache)
         }
 
-        _ = storeCacheData(
+        _ = Cache.storeCacheData(
           cacheData.copy(
             ontologies = cacheData.ontologies + (internalOntologyIri -> updatedOntology),
             subPropertyOfRelations = updatedSubPropertyOfRelations
@@ -2606,33 +2208,6 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
       userCanUpdateOntology <- canUserUpdateOntology(internalOntologyIri, canDeleteOntologyRequest.requestingUser)
       subjectsUsingOntology <- getSubjectsUsingOntology(ontology)
     } yield CanDoResponseV2(userCanUpdateOntology && subjectsUsingOntology.isEmpty)
-  }
-
-  /**
-    * Gets the set of subjects that refer to an ontology or its entities.
-    *
-    * @param ontology the ontology.
-    * @return the set of subjects that refer to the ontology or its entities.
-    */
-  private def getSubjectsUsingOntology(ontology: ReadOntologyV2): Future[Set[IRI]] = {
-    for {
-      isOntologyUsedSparql <- Future(
-        org.knora.webapi.messages.twirl.queries.sparql.v2.txt
-          .isOntologyUsed(
-            triplestore = settings.triplestoreType,
-            ontologyNamedGraphIri = ontology.ontologyMetadata.ontologyIri,
-            classIris = ontology.classes.keySet,
-            propertyIris = ontology.properties.keySet
-          )
-          .toString())
-
-      isOntologyUsedResponse: SparqlSelectResult <- (storeManager ? SparqlSelectRequest(isOntologyUsedSparql))
-        .mapTo[SparqlSelectResult]
-
-      subjects = isOntologyUsedResponse.results.bindings.map { row =>
-        row.rowMap("s")
-      }.toSet
-    } yield subjects
   }
 
   private def deleteOntology(deleteOntologyRequest: DeleteOntologyRequestV2): Future[SuccessResponseV2] = {
@@ -2700,7 +2275,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
               subClass -> baseClasses.filterNot(_.getOntologyFromEntity == internalOntologyIri)
           }
 
-        updatedSuperClassOfRelations = calculateSuperClassOfRelations(updatedSubClassOfRelations)
+        updatedSuperClassOfRelations = Ontology.calculateSuperClassOfRelations(updatedSubClassOfRelations)
 
         updatedSubPropertyOfRelations = cacheData.subPropertyOfRelations
           .filterNot {
@@ -2838,7 +2413,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
         }
 
         // Check the property's salsah-gui:guiElement and salsah-gui:guiAttribute.
-        _ = validateGuiAttributes(
+        _ = Ontology.validateGuiAttributes(
           propertyInfoContent = internalPropertyDef,
           allGuiAttributeDefinitions = cacheData.guiAttributeDefinitions,
           errorFun = { msg: String =>
@@ -2867,7 +2442,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
           _.requireIriObject(throw BadRequestException("Invalid knora-api:subjectType")))
 
         _ = maybeSubjectClassConstraint.foreach { subjectClassConstraint =>
-          if (!isKnoraInternalResourceClass(subjectClassConstraint, cacheData)) {
+          if (!Ontology.isKnoraInternalResourceClass(subjectClassConstraint, cacheData)) {
             throw BadRequestException(
               s"Invalid subject class constraint: ${subjectClassConstraint.toOntologySchema(ApiV2Complex)}")
           }
@@ -2892,7 +2467,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
 
         _ = maybeSubjectClassConstraint match {
           case Some(subjectClassConstraint) =>
-            checkPropertyConstraint(
+            Cache.checkPropertyConstraint(
               cacheData = cacheData,
               internalPropertyIri = internalPropertyIri,
               constraintPredicateIri = OntologyConstants.KnoraBase.SubjectClassConstraint.toSmartIri,
@@ -2909,7 +2484,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
 
         // Check that the object class is a subclass of the object classes of the base properties.
 
-        _ = checkPropertyConstraint(
+        _ = Cache.checkPropertyConstraint(
           cacheData = cacheData,
           internalPropertyIri = internalPropertyIri,
           constraintPredicateIri = OntologyConstants.KnoraBase.ObjectClassConstraint.toSmartIri,
@@ -2922,7 +2497,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
         )
 
         // Check that the property definition doesn't refer to any non-shared ontologies in other projects.
-        _ = checkOntologyReferencesInPropertyDef(
+        _ = Cache.checkOntologyReferencesInPropertyDef(
           ontologyCacheData = cacheData,
           propertyDef = internalPropertyDef,
           errorFun = { msg: String =>
@@ -2950,7 +2525,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
 
         // Check that the ontology's last modification date was updated.
 
-        _ <- checkOntologyLastModificationDateAfterUpdate(
+        _ <- Ontology.checkOntologyLastModificationDateAfterUpdate(
           internalOntologyIri = internalOntologyIri,
           expectedLastModificationDate = currentTime,
           featureFactoryConfig = createPropertyRequest.featureFactoryConfig
@@ -2959,7 +2534,9 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
         // Check that the data that was saved corresponds to the data that was submitted. To make this comparison,
         // we have to undo the SPARQL-escaping of the input.
 
-        loadedPropertyDef <- loadPropertyDefinition(
+        loadedPropertyDef <- Ontology.loadPropertyDefinition(
+          settings,
+          storeManager,
           propertyIri = internalPropertyIri,
           featureFactoryConfig = createPropertyRequest.featureFactoryConfig
         )
@@ -2973,7 +2550,9 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
 
         maybeLoadedLinkValuePropertyDefFuture: Option[Future[PropertyInfoContentV2]] = maybeLinkValuePropertyDef.map {
           linkValuePropertyDef =>
-            loadPropertyDefinition(
+            Ontology.loadPropertyDefinition(
+              settings,
+              storeManager,
               propertyIri = linkValuePropertyDef.propertyIri,
               featureFactoryConfig = createPropertyRequest.featureFactoryConfig
             )
@@ -3020,7 +2599,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
           properties = ontology.properties ++ maybeLinkValuePropertyCacheEntry + (internalPropertyIri -> readPropertyInfo)
         )
 
-        _ = storeCacheData(
+        _ = Cache.storeCacheData(
           cacheData.copy(
             ontologies = cacheData.ontologies + (internalOntologyIri -> updatedOntology),
             subPropertyOfRelations = cacheData.subPropertyOfRelations + (internalPropertyIri -> allKnoraSuperPropertyIris)
@@ -3042,7 +2621,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
       externalPropertyIri = createPropertyRequest.propertyInfoContent.propertyIri
       externalOntologyIri = externalPropertyIri.getOntologyFromEntity
 
-      _ <- checkOntologyAndEntityIrisForUpdate(
+      _ <- Ontology.checkOntologyAndEntityIrisForUpdate(
         externalOntologyIri = externalOntologyIri,
         externalEntityIri = externalPropertyIri,
         requestingUser = requestingUser
@@ -3083,7 +2662,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
           throw NotFoundException(s"Property ${changePropertyGuiElementRequest.propertyIri} not found"))
 
         // Check that the ontology exists and has not been updated by another user since the client last read it.
-        _ <- checkOntologyLastModificationDateBeforeUpdate(
+        _ <- Ontology.checkOntologyLastModificationDateBeforeUpdate(
           internalOntologyIri = internalOntologyIri,
           expectedLastModificationDate = changePropertyGuiElementRequest.lastModificationDate,
           featureFactoryConfig = changePropertyGuiElementRequest.featureFactoryConfig
@@ -3123,7 +2702,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
 
         // Check that the ontology's last modification date was updated.
 
-        _ <- checkOntologyLastModificationDateAfterUpdate(
+        _ <- Ontology.checkOntologyLastModificationDateAfterUpdate(
           internalOntologyIri = internalOntologyIri,
           expectedLastModificationDate = currentTime,
           featureFactoryConfig = changePropertyGuiElementRequest.featureFactoryConfig
@@ -3132,7 +2711,9 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
         // Check that the data that was saved corresponds to the data that was submitted. To make this comparison,
         // we have to undo the SPARQL-escaping of the input.
 
-        loadedPropertyDef <- loadPropertyDefinition(
+        loadedPropertyDef <- Ontology.loadPropertyDefinition(
+          settings,
+          storeManager,
           propertyIri = internalPropertyIri,
           featureFactoryConfig = changePropertyGuiElementRequest.featureFactoryConfig
         )
@@ -3170,7 +2751,9 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
 
         maybeLoadedLinkValuePropertyDefFuture: Option[Future[PropertyInfoContentV2]] = maybeCurrentLinkValueReadPropertyInfo
           .map { linkValueReadPropertyInfo =>
-            loadPropertyDefinition(
+            Ontology.loadPropertyDefinition(
+              settings,
+              storeManager,
               propertyIri = linkValueReadPropertyInfo.entityInfoContent.propertyIri,
               featureFactoryConfig = changePropertyGuiElementRequest.featureFactoryConfig
             )
@@ -3224,7 +2807,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
           properties = ontology.properties ++ maybeLinkValuePropertyCacheEntry + (internalPropertyIri -> newReadPropertyInfo)
         )
 
-        _ = storeCacheData(
+        _ = Cache.storeCacheData(
           cacheData.copy(
             ontologies = cacheData.ontologies + (internalOntologyIri -> updatedOntology)
           ))
@@ -3244,7 +2827,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
       externalPropertyIri = changePropertyGuiElementRequest.propertyIri
       externalOntologyIri = externalPropertyIri.getOntologyFromEntity
 
-      _ <- checkOntologyAndEntityIrisForUpdate(
+      _ <- Ontology.checkOntologyAndEntityIrisForUpdate(
         externalOntologyIri = externalOntologyIri,
         externalEntityIri = externalPropertyIri,
         requestingUser = requestingUser
@@ -3285,7 +2868,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
           throw NotFoundException(s"Property ${changePropertyLabelsOrCommentsRequest.propertyIri} not found"))
 
         // Check that the ontology exists and has not been updated by another user since the client last read it.
-        _ <- checkOntologyLastModificationDateBeforeUpdate(
+        _ <- Ontology.checkOntologyLastModificationDateBeforeUpdate(
           internalOntologyIri = internalOntologyIri,
           expectedLastModificationDate = changePropertyLabelsOrCommentsRequest.lastModificationDate,
           featureFactoryConfig = changePropertyLabelsOrCommentsRequest.featureFactoryConfig
@@ -3325,7 +2908,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
 
         // Check that the ontology's last modification date was updated.
 
-        _ <- checkOntologyLastModificationDateAfterUpdate(
+        _ <- Ontology.checkOntologyLastModificationDateAfterUpdate(
           internalOntologyIri = internalOntologyIri,
           expectedLastModificationDate = currentTime,
           featureFactoryConfig = changePropertyLabelsOrCommentsRequest.featureFactoryConfig
@@ -3334,7 +2917,9 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
         // Check that the data that was saved corresponds to the data that was submitted. To make this comparison,
         // we have to undo the SPARQL-escaping of the input.
 
-        loadedPropertyDef <- loadPropertyDefinition(
+        loadedPropertyDef <- Ontology.loadPropertyDefinition(
+          settings,
+          storeManager,
           propertyIri = internalPropertyIri,
           featureFactoryConfig = changePropertyLabelsOrCommentsRequest.featureFactoryConfig
         )
@@ -3355,7 +2940,9 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
 
         maybeLoadedLinkValuePropertyDefFuture: Option[Future[PropertyInfoContentV2]] = maybeCurrentLinkValueReadPropertyInfo
           .map { linkValueReadPropertyInfo =>
-            loadPropertyDefinition(
+            Ontology.loadPropertyDefinition(
+              settings,
+              storeManager,
               propertyIri = linkValueReadPropertyInfo.entityInfoContent.propertyIri,
               featureFactoryConfig = changePropertyLabelsOrCommentsRequest.featureFactoryConfig
             )
@@ -3405,7 +2992,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
           properties = ontology.properties ++ maybeLinkValuePropertyCacheEntry + (internalPropertyIri -> newReadPropertyInfo)
         )
 
-        _ = storeCacheData(
+        _ = Cache.storeCacheData(
           cacheData.copy(
             ontologies = cacheData.ontologies + (internalOntologyIri -> updatedOntology)
           ))
@@ -3425,7 +3012,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
       externalPropertyIri = changePropertyLabelsOrCommentsRequest.propertyIri
       externalOntologyIri = externalPropertyIri.getOntologyFromEntity
 
-      _ <- checkOntologyAndEntityIrisForUpdate(
+      _ <- Ontology.checkOntologyAndEntityIrisForUpdate(
         externalOntologyIri = externalOntologyIri,
         externalEntityIri = externalPropertyIri,
         requestingUser = requestingUser
@@ -3465,7 +3052,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
           throw NotFoundException(s"Class ${changeClassLabelsOrCommentsRequest.classIri} not found"))
 
         // Check that the ontology exists and has not been updated by another user since the client last read it.
-        _ <- checkOntologyLastModificationDateBeforeUpdate(
+        _ <- Ontology.checkOntologyLastModificationDateBeforeUpdate(
           internalOntologyIri = internalOntologyIri,
           expectedLastModificationDate = changeClassLabelsOrCommentsRequest.lastModificationDate,
           featureFactoryConfig = changeClassLabelsOrCommentsRequest.featureFactoryConfig
@@ -3492,7 +3079,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
 
         // Check that the ontology's last modification date was updated.
 
-        _ <- checkOntologyLastModificationDateAfterUpdate(
+        _ <- Ontology.checkOntologyLastModificationDateAfterUpdate(
           internalOntologyIri = internalOntologyIri,
           expectedLastModificationDate = currentTime,
           featureFactoryConfig = changeClassLabelsOrCommentsRequest.featureFactoryConfig
@@ -3501,7 +3088,9 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
         // Check that the data that was saved corresponds to the data that was submitted. To make this comparison,
         // we have to undo the SPARQL-escaping of the input.
 
-        loadedClassDef: ClassInfoContentV2 <- loadClassDefinition(
+        loadedClassDef: ClassInfoContentV2 <- Ontology.loadClassDefinition(
+          settings,
+          storeManager,
           classIri = internalClassIri,
           featureFactoryConfig = changeClassLabelsOrCommentsRequest.featureFactoryConfig
         )
@@ -3533,7 +3122,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
           classes = ontology.classes + (internalClassIri -> newReadClassInfo)
         )
 
-        _ = storeCacheData(
+        _ = Cache.storeCacheData(
           cacheData.copy(
             ontologies = cacheData.ontologies + (internalOntologyIri -> updatedOntology)
           ))
@@ -3554,7 +3143,7 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
       externalClassIri = changeClassLabelsOrCommentsRequest.classIri
       externalOntologyIri = externalClassIri.getOntologyFromEntity
 
-      _ <- checkOntologyAndEntityIrisForUpdate(
+      _ <- Ontology.checkOntologyAndEntityIrisForUpdate(
         externalOntologyIri = externalOntologyIri,
         externalEntityIri = externalClassIri,
         requestingUser = requestingUser
@@ -3576,876 +3165,4 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
     } yield taskResult
   }
 
-  /**
-    * Before an update of an ontology entity, checks that the entity's external IRI, and that of its ontology,
-    * are valid, and checks that the user has permission to update the ontology.
-    *
-    * @param externalOntologyIri the external IRI of the ontology.
-    * @param externalEntityIri   the external IRI of the entity.
-    * @param requestingUser      the user making the request.
-    */
-  private def checkOntologyAndEntityIrisForUpdate(externalOntologyIri: SmartIri,
-                                                  externalEntityIri: SmartIri,
-                                                  requestingUser: UserADM): Future[Unit] = {
-    for {
-      _ <- checkExternalOntologyIriForUpdate(externalOntologyIri)
-      _ <- checkExternalEntityIriForUpdate(externalEntityIri = externalEntityIri)
-      _ <- checkPermissionsForOntologyUpdate(
-        internalOntologyIri = externalOntologyIri.toOntologySchema(InternalSchema),
-        requestingUser = requestingUser
-      )
-    } yield ()
-  }
-
-  /**
-    * Loads a property definition from the triplestore and converts it to a [[PropertyInfoContentV2]].
-    *
-    * @param propertyIri the IRI of the property to be loaded.
-    * @param featureFactoryConfig the feature factory configuration.
-    * @return a [[PropertyInfoContentV2]] representing the property definition.
-    */
-  private def loadPropertyDefinition(propertyIri: SmartIri,
-                                     featureFactoryConfig: FeatureFactoryConfig): Future[PropertyInfoContentV2] = {
-    for {
-      sparql <- Future(
-        org.knora.webapi.messages.twirl.queries.sparql.v2.txt
-          .getPropertyDefinition(
-            triplestore = settings.triplestoreType,
-            propertyIri = propertyIri
-          )
-          .toString())
-
-      constructResponse <- (storeManager ? SparqlExtendedConstructRequest(
-        sparql = sparql,
-        featureFactoryConfig = featureFactoryConfig
-      )).mapTo[SparqlExtendedConstructResponse]
-    } yield
-      constructResponseToPropertyDefinition(
-        propertyIri = propertyIri,
-        constructResponse = constructResponse
-      )
-  }
-
-  /**
-    * Given a map of predicate IRIs to predicate objects describing an entity, returns a map of smart IRIs to [[PredicateInfoV2]]
-    * objects that can be used to construct an [[EntityInfoContentV2]].
-    *
-    * @param entityDefMap a map of predicate IRIs to predicate objects.
-    * @return a map of smart IRIs to [[PredicateInfoV2]] objects.
-    */
-  private def getEntityPredicatesFromConstructResponse(
-      entityDefMap: Map[SmartIri, Seq[LiteralV2]]): Map[SmartIri, PredicateInfoV2] = {
-    entityDefMap.map {
-      case (predicateIri: SmartIri, predObjs: Seq[LiteralV2]) =>
-        val predicateInfo = PredicateInfoV2(
-          predicateIri = predicateIri,
-          objects = predObjs.map {
-            case IriLiteralV2(iriStr) =>
-              // We use xsd:dateTime in the triplestore (because it is supported in SPARQL), but we return
-              // the more restrictive xsd:dateTimeStamp in the API.
-              if (iriStr == OntologyConstants.Xsd.DateTime) {
-                SmartIriLiteralV2(OntologyConstants.Xsd.DateTimeStamp.toSmartIri)
-              } else {
-                SmartIriLiteralV2(iriStr.toSmartIri)
-              }
-
-            case ontoLiteral: OntologyLiteralV2 => ontoLiteral
-
-            case other =>
-              throw InconsistentRepositoryDataException(s"Predicate $predicateIri has an invalid object: $other")
-          }
-        )
-
-        predicateIri -> predicateInfo
-    }
-  }
-
-  /**
-    * Extracts property definitions from a SPARQL CONSTRUCT response.
-    *
-    * @param propertyIris      the IRIs of the properties to be read.
-    * @param constructResponse the SPARQL construct response to be read.
-    * @return a map of property IRIs to property definitions.
-    */
-  private def constructResponseToPropertyDefinitions(
-      propertyIris: Set[SmartIri],
-      constructResponse: SparqlExtendedConstructResponse): Map[SmartIri, PropertyInfoContentV2] = {
-    propertyIris.map { propertyIri =>
-      propertyIri -> constructResponseToPropertyDefinition(
-        propertyIri = propertyIri,
-        constructResponse = constructResponse
-      )
-    }.toMap
-  }
-
-  /**
-    * Converts a SPARQL CONSTRUCT response to a [[PropertyInfoContentV2]].
-    *
-    * @param propertyIri       the IRI of the property to be read.
-    * @param constructResponse the SPARQL CONSTRUCT response to be read.
-    * @return a [[PropertyInfoContentV2]] representing a property definition.
-    */
-  private def constructResponseToPropertyDefinition(
-      propertyIri: SmartIri,
-      constructResponse: SparqlExtendedConstructResponse): PropertyInfoContentV2 = {
-    // All properties defined in the triplestore must be in Knora ontologies.
-
-    val ontologyIri = propertyIri.getOntologyFromEntity
-
-    if (!ontologyIri.isKnoraOntologyIri) {
-      throw InconsistentRepositoryDataException(s"Property $propertyIri is not in a Knora ontology")
-    }
-
-    val statements = constructResponse.statements
-
-    // Get the statements whose subject is the property.
-    val propertyDefMap: Map[SmartIri, Seq[LiteralV2]] = statements(IriSubjectV2(propertyIri.toString))
-
-    val subPropertyOf: Set[SmartIri] = propertyDefMap.get(OntologyConstants.Rdfs.SubPropertyOf.toSmartIri) match {
-      case Some(baseProperties) =>
-        baseProperties.map {
-          case iriLiteral: IriLiteralV2 => iriLiteral.value.toSmartIri
-          case other                    => throw InconsistentRepositoryDataException(s"Unexpected object for rdfs:subPropertyOf: $other")
-        }.toSet
-
-      case None => Set.empty[SmartIri]
-    }
-
-    val otherPreds: Map[SmartIri, PredicateInfoV2] = getEntityPredicatesFromConstructResponse(
-      propertyDefMap - OntologyConstants.Rdfs.SubPropertyOf.toSmartIri)
-
-    // salsah-gui:guiOrder isn't allowed here.
-    if (otherPreds.contains(OntologyConstants.SalsahGui.GuiOrder.toSmartIri)) {
-      throw InconsistentRepositoryDataException(s"Property $propertyIri contains salsah-gui:guiOrder")
-    }
-
-    val propertyDef = PropertyInfoContentV2(
-      propertyIri = propertyIri,
-      subPropertyOf = subPropertyOf,
-      predicates = otherPreds,
-      ontologySchema = propertyIri.getOntologySchema.get
-    )
-
-    if (!propertyIri.isKnoraBuiltInDefinitionIri && propertyDef.getRdfTypes.contains(
-          OntologyConstants.Owl.TransitiveProperty.toSmartIri)) {
-      throw InconsistentRepositoryDataException(
-        s"Project-specific property $propertyIri cannot be an owl:TransitiveProperty")
-    }
-
-    propertyDef
-  }
-
-  /**
-    * Reads OWL named individuals from a SPARQL CONSTRUCT response.
-    *
-    * @param individualIris    the IRIs of the named individuals to be read.
-    * @param constructResponse the SPARQL CONSTRUCT response.
-    * @return a map of individual IRIs to named individuals.
-    */
-  private def constructResponseToIndividuals(
-      individualIris: Set[SmartIri],
-      constructResponse: SparqlExtendedConstructResponse): Map[SmartIri, IndividualInfoContentV2] = {
-    individualIris.map { individualIri =>
-      individualIri -> constructResponseToIndividual(
-        individualIri = individualIri,
-        constructResponse = constructResponse
-      )
-    }.toMap
-  }
-
-  /**
-    * Reads an OWL named individual from a SPARQL CONSTRUCT response.
-    *
-    * @param individualIri     the IRI of the individual to be read.
-    * @param constructResponse the SPARQL CONSTRUCT response.
-    * @return an [[IndividualInfoContentV2]] representing the named individual.
-    */
-  private def constructResponseToIndividual(
-      individualIri: SmartIri,
-      constructResponse: SparqlExtendedConstructResponse): IndividualInfoContentV2 = {
-    val statements = constructResponse.statements
-
-    // Get the statements whose subject is the individual.
-    val individualMap: Map[SmartIri, Seq[LiteralV2]] = statements(IriSubjectV2(individualIri.toString))
-
-    val predicates: Map[SmartIri, PredicateInfoV2] = getEntityPredicatesFromConstructResponse(individualMap)
-
-    IndividualInfoContentV2(
-      individualIri = individualIri,
-      predicates = predicates,
-      ontologySchema = individualIri.getOntologySchema.get
-    )
-  }
-
-  /**
-    * Loads a class definition from the triplestore and converts it to a [[ClassInfoContentV2]].
-    *
-    * @param classIri the IRI of the class to be loaded.
-    * @param featureFactoryConfig the feature factory configuration.
-    * @return a [[ClassInfoContentV2]] representing the class definition.
-    */
-  private def loadClassDefinition(classIri: SmartIri,
-                                  featureFactoryConfig: FeatureFactoryConfig): Future[ClassInfoContentV2] = {
-    for {
-      sparql <- Future(
-        org.knora.webapi.messages.twirl.queries.sparql.v2.txt
-          .getClassDefinition(
-            triplestore = settings.triplestoreType,
-            classIri = classIri
-          )
-          .toString())
-
-      constructResponse <- (storeManager ? SparqlExtendedConstructRequest(
-        sparql = sparql,
-        featureFactoryConfig = featureFactoryConfig
-      )).mapTo[SparqlExtendedConstructResponse]
-    } yield
-      constructResponseToClassDefinition(
-        classIri = classIri,
-        constructResponse = constructResponse
-      )
-  }
-
-  /**
-    * Extracts class definitions from a SPARQL CONSTRUCT response.
-    *
-    * @param classIris         the IRIs of the classes to be read.
-    * @param constructResponse the SPARQL CONSTRUCT response to be read.
-    * @return a map of class IRIs to class definitions.
-    */
-  private def constructResponseToClassDefinitions(
-      classIris: Set[SmartIri],
-      constructResponse: SparqlExtendedConstructResponse): Map[SmartIri, ClassInfoContentV2] = {
-    classIris.map { classIri =>
-      classIri -> constructResponseToClassDefinition(
-        classIri = classIri,
-        constructResponse = constructResponse
-      )
-    }.toMap
-  }
-
-  /**
-    * Converts a SPARQL CONSTRUCT response to a [[ClassInfoContentV2]].
-    *
-    * @param classIri          the IRI of the class to be read.
-    * @param constructResponse the SPARQL CONSTRUCT response to be read.
-    * @return a [[ClassInfoContentV2]] representing a class definition.
-    */
-  private def constructResponseToClassDefinition(
-      classIri: SmartIri,
-      constructResponse: SparqlExtendedConstructResponse): ClassInfoContentV2 = {
-    // All classes defined in the triplestore must be in Knora ontologies.
-
-    val ontologyIri = classIri.getOntologyFromEntity
-
-    if (!ontologyIri.isKnoraOntologyIri) {
-      throw InconsistentRepositoryDataException(s"Class $classIri is not in a Knora ontology")
-    }
-
-    val statements = constructResponse.statements
-
-    // Get the statements whose subject is the class.
-    val classDefMap: Map[SmartIri, Seq[LiteralV2]] = statements(IriSubjectV2(classIri.toString))
-
-    // Get the IRIs of the class's base classes.
-
-    val subClassOfObjects: Seq[LiteralV2] =
-      classDefMap.getOrElse(OntologyConstants.Rdfs.SubClassOf.toSmartIri, Seq.empty[LiteralV2])
-
-    val subClassOf: Set[SmartIri] = subClassOfObjects.collect {
-      case iriLiteral: IriLiteralV2 => iriLiteral.value.toSmartIri
-    }.toSet
-
-    // Get the blank nodes representing cardinalities.
-
-    val restrictionBlankNodeIDs: Set[BlankNodeLiteralV2] = subClassOfObjects.collect {
-      case blankNodeLiteral: BlankNodeLiteralV2 => blankNodeLiteral
-    }.toSet
-
-    val directCardinalities: Map[SmartIri, KnoraCardinalityInfo] = restrictionBlankNodeIDs.map { blankNodeID =>
-      val blankNode: Map[SmartIri, Seq[LiteralV2]] = statements.getOrElse(
-        BlankNodeSubjectV2(blankNodeID.value),
-        throw InconsistentRepositoryDataException(
-          s"Blank node '${blankNodeID.value}' not found in construct query result")
-      )
-
-      val blankNodeTypeObjs: Seq[LiteralV2] = blankNode.getOrElse(
-        OntologyConstants.Rdf.Type.toSmartIri,
-        throw InconsistentRepositoryDataException(s"Blank node '${blankNodeID.value}' has no rdf:type"))
-
-      blankNodeTypeObjs match {
-        case Seq(IriLiteralV2(OntologyConstants.Owl.Restriction)) => ()
-        case _ =>
-          throw InconsistentRepositoryDataException(s"Blank node '${blankNodeID.value}' is not an owl:Restriction")
-      }
-
-      val onPropertyObjs: Seq[LiteralV2] = blankNode.getOrElse(
-        OntologyConstants.Owl.OnProperty.toSmartIri,
-        throw InconsistentRepositoryDataException(s"Blank node '${blankNodeID.value}' has no owl:onProperty"))
-
-      val propertyIri: SmartIri = onPropertyObjs match {
-        case Seq(propertyIri: IriLiteralV2) => propertyIri.value.toSmartIri
-        case other                          => throw InconsistentRepositoryDataException(s"Invalid object for owl:onProperty: $other")
-      }
-
-      val owlCardinalityPreds: Set[SmartIri] = blankNode.keySet.filter { predicate =>
-        OntologyConstants.Owl.cardinalityOWLRestrictions(predicate.toString)
-      }
-
-      if (owlCardinalityPreds.size != 1) {
-        throw InconsistentRepositoryDataException(
-          s"Expected one cardinality predicate in blank node '${blankNodeID.value}', got ${owlCardinalityPreds.size}")
-      }
-
-      val owlCardinalityIri = owlCardinalityPreds.head
-
-      val owlCardinalityValue: Int = blankNode(owlCardinalityIri) match {
-        case Seq(IntLiteralV2(intVal)) => intVal
-        case other =>
-          throw InconsistentRepositoryDataException(
-            s"Expected one integer object for predicate $owlCardinalityIri in blank node '${blankNodeID.value}', got $other")
-      }
-
-      val guiOrder: Option[Int] = blankNode.get(OntologyConstants.SalsahGui.GuiOrder.toSmartIri) match {
-        case Some(Seq(IntLiteralV2(intVal))) => Some(intVal)
-        case None                            => None
-        case other =>
-          throw InconsistentRepositoryDataException(
-            s"Expected one integer object for predicate ${OntologyConstants.SalsahGui.GuiOrder} in blank node '${blankNodeID.value}', got $other")
-      }
-
-      // salsah-gui:guiElement and salsah-gui:guiAttribute aren't allowed here.
-
-      if (blankNode.contains(OntologyConstants.SalsahGui.GuiElementProp.toSmartIri)) {
-        throw InconsistentRepositoryDataException(
-          s"Class $classIri contains salsah-gui:guiElement in an owl:Restriction")
-      }
-
-      if (blankNode.contains(OntologyConstants.SalsahGui.GuiAttribute.toSmartIri)) {
-        throw InconsistentRepositoryDataException(
-          s"Class $classIri contains salsah-gui:guiAttribute in an owl:Restriction")
-      }
-
-      propertyIri -> Cardinality.owlCardinality2KnoraCardinality(
-        propertyIri = propertyIri.toString,
-        owlCardinality = Cardinality.OwlCardinalityInfo(
-          owlCardinalityIri = owlCardinalityIri.toString,
-          owlCardinalityValue = owlCardinalityValue,
-          guiOrder = guiOrder
-        )
-      )
-    }.toMap
-
-    // Get any other predicates of the class.
-
-    val otherPreds: Map[SmartIri, PredicateInfoV2] = getEntityPredicatesFromConstructResponse(
-      classDefMap - OntologyConstants.Rdfs.SubClassOf.toSmartIri)
-
-    ClassInfoContentV2(
-      classIri = classIri,
-      subClassOf = subClassOf,
-      predicates = otherPreds,
-      directCardinalities = directCardinalities,
-      ontologySchema = classIri.getOntologySchema.get
-    )
-  }
-
-  /**
-    * Checks that a property's `knora-base:subjectClassConstraint` or `knora-base:objectClassConstraint` is compatible with (i.e. a subclass of)
-    * the ones in all its base properties.
-    *
-    * @param internalPropertyIri        the internal IRI of the property to be checked.
-    * @param constraintPredicateIri     the internal IRI of the constraint, i.e. `knora-base:subjectClassConstraint` or `knora-base:objectClassConstraint`.
-    * @param constraintValueToBeChecked the constraint value to be checked.
-    * @param allSuperPropertyIris       the IRIs of all the base properties of the property, including indirect base properties and the property itself.
-    * @param errorSchema                the ontology schema to be used for error messages.
-    * @param errorFun                   a function that throws an exception. It will be called with an error message argument if the property constraint is invalid.
-    */
-  private def checkPropertyConstraint(cacheData: OntologyCacheData,
-                                      internalPropertyIri: SmartIri,
-                                      constraintPredicateIri: SmartIri,
-                                      constraintValueToBeChecked: SmartIri,
-                                      allSuperPropertyIris: Set[SmartIri],
-                                      errorSchema: OntologySchema,
-                                      errorFun: String => Nothing): Unit = {
-    // The property constraint value must be a Knora class, or one of a limited set of classes defined in OWL.
-    val superClassesOfConstraintValueToBeChecked: Set[SmartIri] =
-      if (OntologyConstants.Owl.ClassesThatCanBeKnoraClassConstraints.contains(constraintValueToBeChecked.toString)) {
-        Set(constraintValueToBeChecked)
-      } else {
-        cacheData.subClassOfRelations
-          .getOrElse(
-            constraintValueToBeChecked,
-            errorFun(
-              s"Property ${internalPropertyIri.toOntologySchema(errorSchema)} cannot have a ${constraintPredicateIri
-                .toOntologySchema(errorSchema)} of " +
-                s"${constraintValueToBeChecked.toOntologySchema(errorSchema)}")
-          )
-          .toSet
-      }
-
-    // Get the definitions of all the Knora superproperties of the property.
-    val superPropertyInfos: Set[ReadPropertyInfoV2] = (allSuperPropertyIris - internalPropertyIri).collect {
-      case superPropertyIri if superPropertyIri.isKnoraDefinitionIri =>
-        cacheData
-          .ontologies(superPropertyIri.getOntologyFromEntity)
-          .properties
-          .getOrElse(
-            superPropertyIri,
-            errorFun(
-              s"Property ${internalPropertyIri.toOntologySchema(errorSchema)} is a subproperty of $superPropertyIri, which is undefined")
-          )
-    }
-
-    // For each superproperty definition, get the value of the specified constraint in that definition, if any. Here we
-    // make a map of superproperty IRIs to superproperty constraint values.
-    val superPropertyConstraintValues: Map[SmartIri, SmartIri] = superPropertyInfos.flatMap { superPropertyInfo =>
-      superPropertyInfo.entityInfoContent.predicates
-        .get(constraintPredicateIri)
-        .map(_.requireIriObject(throw InconsistentRepositoryDataException(
-          s"Property ${superPropertyInfo.entityInfoContent.propertyIri} has an invalid object for $constraintPredicateIri")))
-        .map { superPropertyConstraintValue =>
-          superPropertyInfo.entityInfoContent.propertyIri -> superPropertyConstraintValue
-        }
-    }.toMap
-
-    // Check that the constraint value to be checked is a subclass of the constraint value in every superproperty.
-
-    superPropertyConstraintValues.foreach {
-      case (superPropertyIri, superPropertyConstraintValue) =>
-        if (!superClassesOfConstraintValueToBeChecked.contains(superPropertyConstraintValue)) {
-          errorFun(
-            s"Property ${internalPropertyIri.toOntologySchema(errorSchema)} cannot have a ${constraintPredicateIri
-              .toOntologySchema(errorSchema)} of " +
-              s"${constraintValueToBeChecked.toOntologySchema(errorSchema)}, because that is not a subclass of " +
-              s"${superPropertyConstraintValue.toOntologySchema(errorSchema)}, which is the ${constraintPredicateIri
-                .toOntologySchema(errorSchema)} of " +
-              s"a base property, ${superPropertyIri.toOntologySchema(errorSchema)}")
-        }
-    }
-  }
-
-  /**
-    * Checks the last modification date of an ontology before an update.
-    *
-    * @param internalOntologyIri          the internal IRI of the ontology.
-    * @param expectedLastModificationDate the last modification date that should now be attached to the ontology.
-    * @param featureFactoryConfig the feature factory configuration.
-    * @return a failed Future if the expected last modification date is not found.
-    */
-  private def checkOntologyLastModificationDateBeforeUpdate(
-      internalOntologyIri: SmartIri,
-      expectedLastModificationDate: Instant,
-      featureFactoryConfig: FeatureFactoryConfig): Future[Unit] = {
-    checkOntologyLastModificationDate(
-      internalOntologyIri = internalOntologyIri,
-      expectedLastModificationDate = expectedLastModificationDate,
-      featureFactoryConfig = featureFactoryConfig,
-      errorFun = throw EditConflictException(
-        s"Ontology ${internalOntologyIri.toOntologySchema(ApiV2Complex)} has been modified by another user, please reload it and try again.")
-    )
-  }
-
-  /**
-    * Checks the last modification date of an ontology after an update.
-    *
-    * @param internalOntologyIri          the internal IRI of the ontology.
-    * @param expectedLastModificationDate the last modification date that should now be attached to the ontology.
-    * @param featureFactoryConfig the feature factory configuration.
-    * @return a failed Future if the expected last modification date is not found.
-    */
-  private def checkOntologyLastModificationDateAfterUpdate(internalOntologyIri: SmartIri,
-                                                           expectedLastModificationDate: Instant,
-                                                           featureFactoryConfig: FeatureFactoryConfig): Future[Unit] = {
-    checkOntologyLastModificationDate(
-      internalOntologyIri = internalOntologyIri,
-      expectedLastModificationDate = expectedLastModificationDate,
-      featureFactoryConfig = featureFactoryConfig,
-      errorFun = throw UpdateNotPerformedException(
-        s"Ontology ${internalOntologyIri.toOntologySchema(ApiV2Complex)} was not updated. Please report this as a possible bug.")
-    )
-  }
-
-  /**
-    * Checks the last modification date of an ontology.
-    *
-    * @param internalOntologyIri          the internal IRI of the ontology.
-    * @param expectedLastModificationDate the last modification date that the ontology is expected to have.
-    * @param featureFactoryConfig the feature factory configuration.
-    * @param errorFun                     a function that throws an exception. It will be called if the expected last modification date is not found.
-    * @return a failed Future if the expected last modification date is not found.
-    */
-  private def checkOntologyLastModificationDate(internalOntologyIri: SmartIri,
-                                                expectedLastModificationDate: Instant,
-                                                featureFactoryConfig: FeatureFactoryConfig,
-                                                errorFun: => Nothing): Future[Unit] = {
-    for {
-      existingOntologyMetadata: Option[OntologyMetadataV2] <- loadOntologyMetadata(
-        internalOntologyIri = internalOntologyIri,
-        featureFactoryConfig = featureFactoryConfig
-      )
-
-      _ = existingOntologyMetadata match {
-        case Some(metadata) =>
-          metadata.lastModificationDate match {
-            case Some(lastModificationDate) =>
-              if (lastModificationDate != expectedLastModificationDate) {
-                errorFun
-              }
-
-            case None =>
-              throw InconsistentRepositoryDataException(
-                s"Ontology $internalOntologyIri has no ${OntologyConstants.KnoraBase.LastModificationDate}")
-          }
-
-        case None =>
-          throw NotFoundException(
-            s"Ontology $internalOntologyIri (corresponding to ${internalOntologyIri.toOntologySchema(ApiV2Complex)}) not found")
-      }
-    } yield ()
-  }
-
-  /**
-    * Checks whether the requesting user has permission to update an ontology.
-    *
-    * @param internalOntologyIri the internal IRI of the ontology.
-    * @param requestingUser      the user making the request.
-    * @return `true` if the user has permission to update the ontology
-    */
-  private def canUserUpdateOntology(internalOntologyIri: SmartIri, requestingUser: UserADM): Future[Boolean] = {
-    for {
-      cacheData <- getCacheData
-
-      projectIri = cacheData.ontologies
-        .getOrElse(
-          internalOntologyIri,
-          throw NotFoundException(s"Ontology ${internalOntologyIri.toOntologySchema(ApiV2Complex)} not found")
-        )
-        .ontologyMetadata
-        .projectIri
-        .get
-    } yield requestingUser.permissions.isProjectAdmin(projectIri.toString) || requestingUser.permissions.isSystemAdmin
-  }
-
-  /**
-    * Throws an exception if the requesting user does not have permission to update an ontology.
-    *
-    * @param internalOntologyIri the internal IRI of the ontology.
-    * @param requestingUser      the user making the request.
-    * @return the project IRI.
-    */
-  private def checkPermissionsForOntologyUpdate(internalOntologyIri: SmartIri,
-                                                requestingUser: UserADM): Future[SmartIri] = {
-    for {
-      cacheData <- getCacheData
-
-      projectIri = cacheData.ontologies
-        .getOrElse(
-          internalOntologyIri,
-          throw NotFoundException(s"Ontology ${internalOntologyIri.toOntologySchema(ApiV2Complex)} not found")
-        )
-        .ontologyMetadata
-        .projectIri
-        .get
-
-      _ = if (!requestingUser.permissions.isProjectAdmin(projectIri.toString) && !requestingUser.permissions.isSystemAdmin) {
-        // not a project or system admin
-        throw ForbiddenException("Ontologies can be modified only by a project or system admin.")
-      }
-
-    } yield projectIri
-  }
-
-  /**
-    * Checks whether an ontology IRI is valid for an update.
-    *
-    * @param externalOntologyIri the external IRI of the ontology.
-    * @return a failed Future if the IRI is not valid for an update.
-    */
-  private def checkExternalOntologyIriForUpdate(externalOntologyIri: SmartIri): Future[Unit] = {
-    if (!externalOntologyIri.isKnoraOntologyIri) {
-      FastFuture.failed(throw BadRequestException(s"Invalid ontology IRI for request: $externalOntologyIri}"))
-    } else if (!externalOntologyIri.getOntologySchema.contains(ApiV2Complex)) {
-      FastFuture.failed(throw BadRequestException(s"Invalid ontology schema for request: $externalOntologyIri"))
-    } else if (externalOntologyIri.isKnoraBuiltInDefinitionIri) {
-      FastFuture.failed(
-        throw BadRequestException(s"Ontology $externalOntologyIri cannot be modified via the Knora API"))
-    } else {
-      FastFuture.successful(())
-    }
-  }
-
-  /**
-    * Checks whether an entity IRI is valid for an update.
-    *
-    * @param externalEntityIri the external IRI of the entity.
-    * @return a failed Future if the entity IRI is not valid for an update, or is not from the specified ontology.
-    */
-  private def checkExternalEntityIriForUpdate(externalEntityIri: SmartIri): Future[Unit] = {
-    if (!externalEntityIri.isKnoraApiV2EntityIri) {
-      FastFuture.failed(throw BadRequestException(s"Invalid entity IRI for request: $externalEntityIri"))
-    } else if (!externalEntityIri.getOntologySchema.contains(ApiV2Complex)) {
-      FastFuture.failed(throw BadRequestException(s"Invalid ontology schema for request: $externalEntityIri"))
-    } else {
-      FastFuture.successful(())
-    }
-  }
-
-  /**
-    * Given the definition of a link property, returns the definition of the corresponding link value property.
-    *
-    * @param internalPropertyDef the definition of the the link property, in the internal schema.
-    * @return the definition of the corresponding link value property.
-    */
-  private def linkPropertyDefToLinkValuePropertyDef(
-      internalPropertyDef: PropertyInfoContentV2): PropertyInfoContentV2 = {
-    val linkValuePropIri = internalPropertyDef.propertyIri.fromLinkPropToLinkValueProp
-
-    val newPredicates
-      : Map[SmartIri, PredicateInfoV2] = (internalPropertyDef.predicates - OntologyConstants.KnoraBase.ObjectClassConstraint.toSmartIri) +
-      (OntologyConstants.KnoraBase.ObjectClassConstraint.toSmartIri -> PredicateInfoV2(
-        predicateIri = OntologyConstants.KnoraBase.ObjectClassConstraint.toSmartIri,
-        objects = Seq(SmartIriLiteralV2(OntologyConstants.KnoraBase.LinkValue.toSmartIri))
-      ))
-
-    internalPropertyDef.copy(
-      propertyIri = linkValuePropIri,
-      predicates = newPredicates,
-      subPropertyOf = Set(OntologyConstants.KnoraBase.HasLinkToValue.toSmartIri)
-    )
-  }
-
-  /**
-    * Given the cardinalities directly defined on a given class, and the cardinalities that it could inherit (directly
-    * or indirectly) from its base classes, combines the two, filtering out the base class cardinalities ones that are overridden
-    * by cardinalities defined directly on the given class. Checks that if a directly defined cardinality overrides an inheritable one,
-    * the directly defined one is at least as restrictive as the inheritable one.
-    *
-    * @param classIri                  the class IRI.
-    * @param thisClassCardinalities    the cardinalities directly defined on a given resource class.
-    * @param inheritableCardinalities  the cardinalities that the given resource class could inherit from its base classes.
-    * @param allSubPropertyOfRelations a map in which each property IRI points to the full set of its base properties.
-    * @param errorSchema               the ontology schema to be used in error messages.
-    * @param errorFun                  a function that throws an exception. It will be called with an error message argument if the cardinalities are invalid.
-    * @return a map in which each key is the IRI of a property that has a cardinality in the resource class (or that it inherits
-    *         from its base classes), and each value is the cardinality on the property.
-    */
-  private def overrideCardinalities(classIri: SmartIri,
-                                    thisClassCardinalities: Map[SmartIri, KnoraCardinalityInfo],
-                                    inheritableCardinalities: Map[SmartIri, KnoraCardinalityInfo],
-                                    allSubPropertyOfRelations: Map[SmartIri, Set[SmartIri]],
-                                    errorSchema: OntologySchema,
-                                    errorFun: String => Nothing): Map[SmartIri, KnoraCardinalityInfo] = {
-    // A map of directly defined properties to the base class properties they can override.
-    val overrides: Map[SmartIri, Set[SmartIri]] = thisClassCardinalities.map {
-      case (thisClassProp, thisClassCardinality) =>
-        // For each directly defined cardinality, get its base properties, if available.
-        // If the class has a cardinality for a non-Knora property like rdfs:label (which can happen only
-        // if it's a built-in class), we won't have any information about the base properties of that property.
-        val basePropsOfThisClassProp: Set[SmartIri] =
-          allSubPropertyOfRelations.getOrElse(thisClassProp, Set.empty[SmartIri])
-
-        val overriddenBaseProps: Set[SmartIri] = inheritableCardinalities.foldLeft(Set.empty[SmartIri]) {
-          case (acc, (baseClassProp, baseClassCardinality)) =>
-            // Can the directly defined cardinality override the inheritable one?
-            if (thisClassProp == baseClassProp || basePropsOfThisClassProp.contains(baseClassProp)) {
-              // Yes. Is the directly defined one at least as restrictive as the inheritable one?
-
-              if (!Cardinality.isCompatible(directCardinality = thisClassCardinality.cardinality,
-                                            inheritableCardinality = baseClassCardinality.cardinality)) {
-                // No. Throw an exception.
-                errorFun(s"In class <${classIri.toOntologySchema(errorSchema)}>, the directly defined cardinality ${thisClassCardinality.cardinality} on ${thisClassProp.toOntologySchema(
-                  errorSchema)} is not compatible with the inherited cardinality ${baseClassCardinality.cardinality} on ${baseClassProp
-                  .toOntologySchema(errorSchema)}, because it is less restrictive")
-              } else {
-                // Yes. Filter out the inheritable one, because the directly defined one overrides it.
-                acc + baseClassProp
-              }
-            } else {
-              // No. Let the class inherit the inheritable cardinality.
-              acc
-            }
-        }
-
-        thisClassProp -> overriddenBaseProps
-    }
-
-    // A map of base class properties to the directly defined properties that can override them.
-    val reverseOverrides: Map[SmartIri, Set[SmartIri]] = overrides.toVector
-      .flatMap {
-        // Unpack the sets to make an association list.
-        case (thisClassProp: SmartIri, baseClassProps: Set[SmartIri]) =>
-          baseClassProps.map { baseClassProp: SmartIri =>
-            thisClassProp -> baseClassProp
-          }
-      }
-      .map {
-        // Reverse the direction of the association list.
-        case (thisClassProp: SmartIri, baseClassProp: SmartIri) =>
-          baseClassProp -> thisClassProp
-      }
-      .groupBy {
-        // Group by base class prop to make a map.
-        case (baseClassProp: SmartIri, _) => baseClassProp
-      }
-      .map {
-        // Make sets of directly defined props.
-        case (baseClassProp: SmartIri, thisClassProps: immutable.Iterable[(SmartIri, SmartIri)]) =>
-          baseClassProp -> thisClassProps.map {
-            case (_, thisClassProp) => thisClassProp
-          }.toSet
-      }
-
-    // Are there any base class properties that are overridden by more than one directly defined property,
-    // and do any of those base properties have cardinalities that could cause conflicts between the cardinalities
-    // on the directly defined cardinalities?
-    reverseOverrides.foreach {
-      case (baseClassProp, thisClassProps) =>
-        if (thisClassProps.size > 1) {
-          val overriddenCardinality: KnoraCardinalityInfo = inheritableCardinalities(baseClassProp)
-
-          if (overriddenCardinality.cardinality == Cardinality.MustHaveOne || overriddenCardinality.cardinality == Cardinality.MayHaveOne) {
-            errorFun(
-              s"In class <${classIri.toOntologySchema(errorSchema)}>, there is more than one cardinality that would override the inherited cardinality $overriddenCardinality on <${baseClassProp
-                .toOntologySchema(errorSchema)}>")
-          }
-        }
-    }
-
-    thisClassCardinalities ++ inheritableCardinalities.filterNot {
-      case (basePropIri, _) => reverseOverrides.contains(basePropIri)
-    }
-  }
-
-  /**
-    * Given all the `rdfs:subClassOf` relations between classes, calculates all the inverse relations.
-    *
-    * @param allSubClassOfRelations all the `rdfs:subClassOf` relations between classes.
-    * @return a map of IRIs of resource classes to sets of the IRIs of their subclasses.
-    */
-  private def calculateSuperClassOfRelations(
-      allSubClassOfRelations: Map[SmartIri, Seq[SmartIri]]): Map[SmartIri, Set[SmartIri]] = {
-    allSubClassOfRelations.toVector
-      .flatMap {
-        case (subClass: SmartIri, baseClasses: Seq[SmartIri]) =>
-          baseClasses.map { baseClass =>
-            baseClass -> subClass
-          }
-      }
-      .groupBy(_._1)
-      .map {
-        case (baseClass: SmartIri, baseClassAndSubClasses: Vector[(SmartIri, SmartIri)]) =>
-          baseClass -> baseClassAndSubClasses.map(_._2).toSet
-      }
-  }
-
-  /**
-    * Given a class loaded from the triplestore, recursively adds its inherited cardinalities to the cardinalities it defines
-    * directly. A cardinality for a subproperty in a subclass overrides a cardinality for a base property in
-    * a base class.
-    *
-    * @param classIri                  the IRI of the class whose properties are to be computed.
-    * @param directSubClassOfRelations a map of the direct `rdfs:subClassOf` relations defined on each class.
-    * @param allSubPropertyOfRelations a map in which each property IRI points to the full set of its base properties.
-    * @param directClassCardinalities  a map of the cardinalities defined directly on each class.
-    * @return a map in which each key is the IRI of a property that has a cardinality in the class (or that it inherits
-    *         from its base classes), and each value is the cardinality on the property.
-    */
-  private def inheritCardinalitiesInLoadedClass(
-      classIri: SmartIri,
-      directSubClassOfRelations: Map[SmartIri, Set[SmartIri]],
-      allSubPropertyOfRelations: Map[SmartIri, Set[SmartIri]],
-      directClassCardinalities: Map[SmartIri, Map[SmartIri, KnoraCardinalityInfo]])
-    : Map[SmartIri, KnoraCardinalityInfo] = {
-    // Recursively get properties that are available to inherit from base classes. If we have no information about
-    // a class, that could mean that it isn't a subclass of knora-base:Resource (e.g. it's something like
-    // foaf:Person), in which case we assume that it has no base classes.
-    val cardinalitiesAvailableToInherit: Map[SmartIri, KnoraCardinalityInfo] = directSubClassOfRelations
-      .getOrElse(classIri, Set.empty[SmartIri])
-      .foldLeft(Map.empty[SmartIri, KnoraCardinalityInfo]) {
-        case (acc: Map[SmartIri, KnoraCardinalityInfo], baseClass: SmartIri) =>
-          acc ++ inheritCardinalitiesInLoadedClass(
-            classIri = baseClass,
-            directSubClassOfRelations = directSubClassOfRelations,
-            allSubPropertyOfRelations = allSubPropertyOfRelations,
-            directClassCardinalities = directClassCardinalities
-          )
-      }
-
-    // Get the properties that have cardinalities defined directly on this class. Again, if we have no information
-    // about a class, we assume that it has no cardinalities.
-    val thisClassCardinalities: Map[SmartIri, KnoraCardinalityInfo] =
-      directClassCardinalities.getOrElse(classIri, Map.empty[SmartIri, KnoraCardinalityInfo])
-
-    // Combine the cardinalities defined directly on this class with the ones that are available to inherit.
-    overrideCardinalities(
-      classIri = classIri,
-      thisClassCardinalities = thisClassCardinalities,
-      inheritableCardinalities = cardinalitiesAvailableToInherit,
-      allSubPropertyOfRelations = allSubPropertyOfRelations,
-      errorSchema = InternalSchema, { msg: String =>
-        throw InconsistentRepositoryDataException(msg)
-      }
-    )
-  }
-
-  /**
-    * Checks whether a class IRI refers to a Knora internal resource class.
-    *
-    * @param classIri the class IRI.
-    * @return `true` if the class IRI refers to a Knora resource class, or `false` if the class
-    *         does not exist or is not a Knora internal resource class.
-    */
-  private def isKnoraInternalResourceClass(classIri: SmartIri, cacheData: OntologyCacheData): Boolean = {
-    classIri.isKnoraInternalEntityIri &&
-    cacheData.ontologies(classIri.getOntologyFromEntity).classes.get(classIri).exists(_.isResourceClass)
-  }
-
-  /**
-    * Checks whether a property is a subproperty of `knora-base:resourceProperty`.
-    *
-    * @param propertyIri the property IRI.
-    * @param cacheData   the ontology cache.
-    * @return `true` if the property is a subproperty of `knora-base:resourceProperty`.
-    */
-  private def isKnoraResourceProperty(propertyIri: SmartIri, cacheData: OntologyCacheData): Boolean = {
-    propertyIri.isKnoraEntityIri &&
-    cacheData.ontologies(propertyIri.getOntologyFromEntity).properties.get(propertyIri).exists(_.isResourceProp)
-  }
-
-  /**
-    * Checks whether a property is a subproperty of `knora-base:hasLinkTo`.
-    *
-    * @param propertyIri the property IRI.
-    * @param cacheData   the ontology cache.
-    * @return `true` if the property is a subproperty of `knora-base:hasLinkTo`.
-    */
-  private def isLinkProp(propertyIri: SmartIri, cacheData: OntologyCacheData): Boolean = {
-    propertyIri.isKnoraEntityIri &&
-    cacheData.ontologies(propertyIri.getOntologyFromEntity).properties.get(propertyIri).exists(_.isLinkProp)
-  }
-
-  /**
-    * Checks whether a property is a subproperty of `knora-base:hasLinkToValue`.
-    *
-    * @param propertyIri the property IRI.
-    * @param cacheData   the ontology cache.
-    * @return `true` if the property is a subproperty of `knora-base:hasLinkToValue`.
-    */
-  private def isLinkValueProp(propertyIri: SmartIri, cacheData: OntologyCacheData): Boolean = {
-    propertyIri.isKnoraEntityIri &&
-    cacheData.ontologies(propertyIri.getOntologyFromEntity).properties.get(propertyIri).exists(_.isLinkValueProp)
-  }
-
-  /**
-    * Checks whether a property is a subproperty of `knora-base:hasFileValue`.
-    *
-    * @param propertyIri the property IRI.
-    * @param cacheData   the ontology cache.
-    * @return `true` if the property is a subproperty of `knora-base:hasFileValue`.
-    */
-  private def isFileValueProp(propertyIri: SmartIri, cacheData: OntologyCacheData): Boolean = {
-    propertyIri.isKnoraEntityIri &&
-    cacheData.ontologies(propertyIri.getOntologyFromEntity).properties.get(propertyIri).exists(_.isFileValueProp)
-  }
 }
