+++ webapi/src/main/scala/org/knora/webapi/responders/v2/ontology/BUILD.bazel
@@ -1,0 +1,39 @@
+package(default_visibility = ["//visibility:public"])
+
+load("@io_bazel_rules_scala//scala:scala.bzl", "scala_library")
+
+scala_library(
+    name = "ontology",
+    srcs = [
+        "Cache.scala",
+        "DeleteCardinalitiesFromClass.scala",
+        "Ontology.scala",
+    ],
+    scalacopts = ["-deprecation"],
+    unused_dependency_checker_mode = "warn",
+    deps = [
+        "//webapi/src/main/scala/org/knora/webapi",
+        "//webapi/src/main/scala/org/knora/webapi/annotation",
+        "//webapi/src/main/scala/org/knora/webapi/core",
+        "//webapi/src/main/scala/org/knora/webapi/exceptions",
+        "//webapi/src/main/scala/org/knora/webapi/feature",
+        "//webapi/src/main/scala/org/knora/webapi/instrumentation",
+        "//webapi/src/main/scala/org/knora/webapi/messages",
+        "//webapi/src/main/scala/org/knora/webapi/settings",
+        "//webapi/src/main/scala/org/knora/webapi/store/cacheservice/settings",
+        "//webapi/src/main/scala/org/knora/webapi/util",
+        "//webapi/src/main/scala/org/knora/webapi/util/cache",
+        "@maven//:com_typesafe_akka_akka_actor_2_13",
+        "@maven//:com_typesafe_akka_akka_http_2_13",
+        "@maven//:com_typesafe_akka_akka_http_core_2_13",
+        "@maven//:com_typesafe_akka_akka_http_spray_json_2_13",
+        "@maven//:com_typesafe_akka_akka_stream_2_13",
+        "@maven//:com_typesafe_play_twirl_api_2_13",
+        "@maven//:com_typesafe_scala_logging_scala_logging_2_13",
+        "@maven//:io_spray_spray_json_2_13",
+        "@maven//:javax_json_javax_json_api",
+        "@maven//:org_scala_lang_modules_scala_xml_2_13",
+        "@maven//:org_slf4j_slf4j_api",
+        "@maven//:org_springframework_security_spring_security_core",
+    ],
+)
